requirements.txt
pymupdf
opencv-python-headless
pytesseract
Pillow
python-dotenv
flask
openpyxl
gunicorn
numpy
pyzbar
